# pkg/swagger/docs/swagger.yaml
definitions:
...
  emplacc-api_api_v1_dto.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  emplacc-api_api_v1_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorDetail'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
...
  emplacc-api_api_v1_dto.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  emplacc-api_api_v1_dto.ResponseMeta:
    properties:
      pagination:
        $ref: '#/definitions/emplacc-api_api_v1_dto.Pagination'
      trace_id:
        type: string
    type: object
...
  emplacc-api_api_v1_dto.TeamResponse:
    properties:
      data:
        $ref: '#/definitions/emplacc-api_api_v1_dto_response.Team'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
  emplacc-api_api_v1_dto.TeamsListResponse:
    properties:
      data:
        $ref: '#/definitions/emplacc-api_api_v1_dto_response.TeamsPage'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
...
  emplacc-api_api_v1_dto_request.AddTeamMember:
    properties:
      specialization:
        type: string
      user_id:
        type: string
    type: object
  emplacc-api_api_v1_dto_request.AddTeamProject:
    properties:
      project_id:
        type: string
    type: object
...
  emplacc-api_api_v1_dto_request.CreateTeam:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  emplacc-api_api_v1_dto_request.UpdateTeam:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  emplacc-api_api_v1_dto_response.Team:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/emplacc-api_api_v1_dto_response.TeamMember'
        type: array
      name:
        type: string
      updated_at:
        type: string
    type: object
  emplacc-api_api_v1_dto_response.TeamMember:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      specialization:
        type: string
      user_id:
        type: string
    type: object
  emplacc-api_api_v1_dto_response.TeamsPage:
    properties:
      teams:
        items:
          $ref: '#/definitions/emplacc-api_api_v1_dto_response.Team'
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /projects/{project_id}/teams:
    get:
      consumes:
      - application/json
      description: Retrieve a list of teams associated with a specific project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.TeamsListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: List Teams By Project
      tags:
      - Teams
...
  /teams:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of teams
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.TeamsListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: List Teams
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: Create a new team
      parameters:
      - description: Team creation payload
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_api_v1_dto_request.CreateTeam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Create Team
      tags:
      - Teams
  /teams/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a team by its ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Delete Team
      tags:
      - Teams
    get:
      consumes:
      - application/json
      description: Retrieve a team by its ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Get Team
      tags:
      - Teams
    patch:
      consumes:
      - application/json
      description: Update an existing team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Team data
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_api_v1_dto_request.UpdateTeam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Update Team
      tags:
      - Teams
  /teams/{id}/members:
    post:
      consumes:
      - application/json
      description: Add a user to a team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Team member data
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_api_v1_dto_request.AddTeamMember'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Add User to Team
      tags:
      - Teams
  /teams/{id}/members/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove a user from a team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Remove User from Team
      tags:
      - Teams
  /teams/{id}/projects:
    post:
      consumes:
      - application/json
      description: Associate a team with a project
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Team-Project association data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_api_v1_dto_request.AddTeamProject'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Add Team to Project
      tags:
      - Teams
...
swagger: "2.0"